Reorganize string utils
