+++ src/utils/string/formatSymbolWithExchange/formatSymbolWithExchange.test.ts
