+++ src/utils/string/formatSymbolWithExchange/formatSymbolWithExchange.ts
