+++ src/utils/string/formatSymbolWithExchange/index.ts
